requirements.txt
pystray
Pillow
pygame
hidapi
pyinstaller